Add files via upload
First, there are some additions that should be made in the pubspec.yaml of the project. That is why I upload this pubspec.yaml so that you add what is missing (qr_flutter mainly I think).
Just copy the content of each file in lib folder into a new file. Make sure to modify the import statements.